fix problem, now ros2 control works as expected

+++ src/diffdrive_sbem/hardware/include/diffdrive_sbem/arduino_comms.hpp
@@ -84,7 +84,7 @@ public:
     std::string response = send_msg("\r");
   }
 
-  void read_encoder_values(int &val_1, int &val_2)
+  void read_encoder_values(float &val_1, float &val_2)
   {
     std::string response = send_msg("e\r");
 
@@ -92,11 +92,14 @@ public:
     size_t del_pos = response.find(delimiter);
     std::string token_1 = response.substr(0, del_pos);
     std::string token_2 = response.substr(del_pos + delimiter.length());
+    
+    // print for debugging
+    //std::cout << "Read encoders: " << token_1 << " " << token_2 << std::endl;
 
-    val_1 = std::atoi(token_1.c_str());
-    val_2 = std::atoi(token_2.c_str());
+    val_1 = std::atof(token_1.c_str());
+    val_2 = std::atof(token_2.c_str());
   }
-  void set_motor_values(int val_1, int val_2)
+  void set_motor_values(float val_1, float val_2)
   {
     std::stringstream ss;
     ss << "m " << val_1 << " " << val_2 << "\r";
