+++ src/diffdrive_sbem/hardware/include/diffdrive_sbem/wheel.hpp
@@ -10,7 +10,7 @@ class Wheel
     public:
 
     std::string name = "";
-    int enc = 0;
+    float enc = 0;
     double cmd = 0;
     double pos = 0;
     double vel = 0;
